functions/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
cryptography==41.0.4
deprecation==2.1.0
filelock==3.12.4
firebase-admin==6.2.0
firebase-functions==0.1.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fsspec==2023.10.0
functions-framework==3.4.0
gast==0.4.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-events==0.11.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.17.3
idna==3.4
instagrapi==2.0.0
instaloader==4.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.6
watchdog==3.0.0
Werkzeug==3.0.0
wrapt==1.14.1